requirements.txt
Flask
requests
wolframalpha
python-telegram-bot
pyTelegramBotAPI